+++ lib/NaiveText/Category.rb
@@ -1,13 +1,15 @@
 class Category
   @@id_counter = 1
 
-  attr_reader :name
-  attr_reader :id
+  attr_reader :name, :id, :weight
+
 
   def initialize(args)
-    @name = args[:name]
+    @name     = args[:name]
     @examples = args[:examples]
-    @id = @@id_counter
+    @weight   = args[:weight] || 1
+    @id       = @@id_counter
+    
     @@id_counter += 1
   end