+++ lib/NaiveText/TextClassifier.rb
@@ -21,6 +21,10 @@ class TextClassifier
 private
   def get_category_for(text)
     probabilities = @calculator.get_probabilities_for(text)
+    @categories.each do |category|
+      probabilities.multiply(category: category, factor: category.weight)
+      puts category.weight
+    end
     probabilities.category_with_max
   end
 end