+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
--Fixed a typo in the interface of TextClassifier propabilities --> probabilities. Deprecated the old version. 
+- Added optional weighting to categories
+
+##[0.4.2] - 2015-11-02
+-Fixed a typo in the interface of TextClassifier propabilities --> probabilities. Deprecated the old version.
 
 ## [0.4.1] - 2015-10-29
 ### Added